Height Pathfinding
Adding in height to tiles. Characters now move up/down tiles. Pathfind
also takes the height of a tile into consideration. A character is not
able to gain access to a tile that is too high up. A character will
"climp" up lower tiles to get to higher ones.

+++ Assets/Tile Data/TDTile.cs
@@ -21,8 +21,8 @@ public class TDTile
 	}
 
 	int _nHValue; // Heuristic i.e. Distance from current node to target node
-	int _nGValue; // Movement Cost
-	int _nFValue; // G + H
+	float _nGValue; // Movement Cost
+	float _nFValue; // G + H
 	TDTile _parentTDTile;
 
 	TYPE _tileType;
@@ -82,7 +82,7 @@ public class TDTile
 		}
 	}	
 
-	public int GValue {
+	public float GValue {
 		get {
 			return _nGValue;
 		}
@@ -91,7 +91,7 @@ public class TDTile
 		}
 	}
 
-	public int FValue {
+	public float FValue {
 		get {
 			return _nFValue;
 		}
